// veaarevna/src/app/models/Torem.ts
import { Product } from './Products';

export class Torem
{

    Business_name: string='';
    torem_address: string='';
    torem_email: string='';
    torem_phone: string='';
    // product:Product;

    // constructor(Business_name, torem_address,torem_email,torem_phone,)
    // {
    //     this.Business_name=Business_name;
    //     this.torem_address=torem_address;
    //     this.torem_email=torem_email;
    //     this.torem_phone=torem_phone;
    // }

    constructor(){}

}